Add piped filter on web ui
